test(RacingCars): RacingCar 객체와 RacingCars 객체 테스트 코드 분리
