+++ src/test/java/racing/RacingCarTest.java
@@ -14,19 +14,6 @@ import racing.view.UserOutput;
 public class RacingCarTest {
 
 	@Test
-	@DisplayName("레이싱 복수 객체 주행 출력 테스트")
-	void printCarsTest() {
-
-		RacingCars cars = new RacingCars(5);
-
-		for (int i = 0; i < 10; i++) {
-			cars.start();
-			UserOutput.printCarLocation(cars.getRacingCars());
-		}
-	}
-
-
-	@Test
 	@DisplayName("레이싱 단일 객체 주행 출력 테스트")
 	void printCarTest() {
 
@@ -36,27 +23,6 @@ public class RacingCarTest {
 			car.racing();
 			UserOutput.printCarLocation(car.getDistance());
 		}
-	}
-
-	@Test
-	@DisplayName("자동차 여러대 객체 생성 및 주행 테스트")
-	void createCarsRacingTest() {
-
-		RacingCars cars = new RacingCars(5);
-
-		for (int i = 0; i < 10; i++) {
-			cars.start();
-		}
-	}
-
-	@Test
-	@DisplayName("랜덤 상수 생성 테스트")
-	void randomTest() {
-
-		int number = new Random().nextInt(10);
-
-		System.out.println(number);
-		assertThat(number).isLessThan(10);
 	}
 
 	@Test
@@ -71,5 +37,15 @@ public class RacingCarTest {
 		}
 
 		assertThat(car.getDistance()).isLessThan(10);
+	}
+
+	@Test
+	@DisplayName("랜덤 상수 생성 테스트")
+	void randomTest() {
+
+		int number = new Random().nextInt(10);
+
+		System.out.println(number);
+		assertThat(number).isLessThan(10);
 	}
 }
